Update app configuration and dependencies for improved compatibility and iOS settings
This commit modifies the app.json file to change the scheme to "reptracker" and removes deprecated properties for better compatibility. The eas.json file is updated to include iOS-specific submission settings, enhancing the deployment process. Additionally, the package.json and package-lock.json files are updated to reflect the new version of the Expo SDK, ensuring alignment with the latest features and improvements. A new key is added to the Info.plist to specify non-exempt encryption usage, further refining the app's compliance. These changes contribute to maintaining adherence to GymLog's design standards and improving the application's functionality.